requirements.txt
Django == 4.2.5
firebase_admin == 6.2.0
djangorestframework == 3.14.0